Create pitchPredictorForMLR.py
Quick and short ML program to predict what number a certain pitcher will pick based on certain features from earlier picks.

This is of course not even close to being 100p precise, but it indicates some sort of guess.

It measure variance and lets you chose how many features you want to use. It furthermore allows you to chose to find the best parameters for the regression, or to chose standard parameters.