Add files via upload
This is a python script you can run to send you a notification when a bib is available for either cph half or marathon. Currently it is set up for the half, but you can just change the website to scrape. 

Add your phonenumber to the recipients list, or add all the numbers you want and then run the script from the terminal. Remember to check if you have all modules installed. 

Currently only works on OSX as it uses AppleScript to text the notification. Feel free to change it however you like.